PROJECT_FOLDER/requirements.txt
streamlit
pandas
numpy
pyodbc
psycopg2-binary
scikit-learn
matplotlib
seaborn
plotly
opencv-python
pyzbar
python-barcode